i9500: audio: mixer_paths.xml: Fix the echo problem
The echo problem was caused by the main mic being "too" strong, and causing in to catch
the earpiece sound, which made the other side to hear themselves more than once.
Reducing it to this level corrects the "overhear", but not harming the sound call quality,
as the main mic is much stronger than we thought.